move image analysis to Global components
